Support db, dynamic schema and partition key
