+++ example/db_support/clean_data.py
@@ -1,8 +1,9 @@
+import os
 from pymilvus import (
     connections,
+    db,
     utility,
 )
-import os
 
 fmt = "\n=== {:30} ===\n"
 
@@ -13,14 +14,16 @@ if host == None:
 print(fmt.format(f"Milvus host: {host}"))
 connections.connect("default", host=host, port="19530")
 
-print(fmt.format("Drop collection `hello_milvus`"))
+db.using_database(db_name="db1")
+print(fmt.format("Drop collection `db1.hello_milvus`"))
 utility.drop_collection("hello_milvus")
 
-print(fmt.format("Drop collection `hello_milvus2`"))
-utility.drop_collection("hello_milvus2")
-
-print(fmt.format(f"Drop collection `hello_milvus_recover`"))
+print(fmt.format(f"Drop collection `db1.hello_milvus_recover`"))
 utility.drop_collection("hello_milvus_recover")
 
-print(fmt.format(f"Drop collection `hello_milvus2_recover`"))
+db.using_database(db_name="db2")
+print(fmt.format("Drop collection `db2.hello_milvus2`"))
+utility.drop_collection("hello_milvus2")
+
+print(fmt.format(f"Drop collection `db2.hello_milvus2_recover`"))
 utility.drop_collection("hello_milvus2_recover")
